提交文档变化。

+++ server/ResourceServer/FileInfo.h
@@ -3,14 +3,14 @@
 
 class FileInfo {
 public:
-	FileInfo(int seq = 0, std::string name = "", int total_size = 0,
-		int trans_size = 0, std::string file_path_str = "")
+	FileInfo(int seq = 0, std::string name = "", int64_t total_size = 0,
+		int64_t trans_size = 0, std::string file_path_str = "")
 		:_seq(seq), _name(name), _total_size(total_size),
 		_trans_size(trans_size), _file_path_str(file_path_str) {}
 	int _seq;
 	std::string _name;
-	int _total_size;
-	int _trans_size;
+	int64_t _total_size;
+	int64_t _trans_size;
 	std::string _file_path_str;
 };
